+++ addons/account/static/src/scss/section_and_note_backend.scss
@@ -27,3 +27,9 @@ table.o_section_and_note_list_view tr.o_data_row {
         }
     }
 }
+
+.o_field_section_and_note_text {
+    > span {
+        white-space: pre-wrap;
+    }
+}
