+++ addons/account/static/src/xml/section_and_note_fields_backend.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<templates>
+    <t t-name="account.sectionAndNoteFields" t-inherit="web.ListRenderer.RecordRow" t-inherit-mode="extension" owl="1">
+        <xpath expr="//td[@class='o_data_cell']" position="attributes">
+            <attribute name="t-att-colspan">allowSectionsAndNotes ? getCellColspan(column, record) : 1</attribute>
+        </xpath>
+    </t>
+    <t t-name="account.sectionAndNoteTable" t-inherit="web.ListRenderer" t-inherit-mode="extension" owl="1">
+        <xpath expr="//table" position="attributes">
+            <attribute name="t-att-class">allowSectionsAndNotes ? 'o_section_and_note_list_view' : ''</attribute>
+        </xpath>
+    </t>
+</templates>
